#8 Extracted new code into a separate file
